refactor: moving constants.ts
